refactor: dedup code for handling service context in rest test app
